Refactor plugin logging
