--- src/platform/plugins/shared/saved_objects_management/server/types.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface SavedObjectsManagementPluginSetup {}
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface SavedObjectsManagementPluginStart {}
-
-export type {
-  SavedObjectMetadata,
-  SavedObjectWithMetadata,
-  SavedObjectRelationKind,
-  SavedObjectRelation,
-  SavedObjectInvalidRelation,
-} from '../common';
